Cadastro finalizado/requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
click==8.2.1
